add logcat fragment
