+++ src/main/res/layout/logcat_viewer_activity_logcat.xml
@@ -1,36 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:orientation="vertical"
-    android:id="@+id/root"
-    android:fitsSystemWindows="true"
-    tools:context=".LogcatActivity"
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/container"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <androidx.appcompat.widget.Toolbar
-        android:id="@+id/toolbar"
-        android:layout_width="match_parent"
-        android:layout_height="?actionBarSize"
-        app:navigationIcon="?attr/homeAsUpIndicator"
-        app:title=""
-        >
-
-        <Spinner
-            android:id="@+id/spinner"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            tools:listitem="@layout/logcat_viewer_item_logcat_dropdown"
-            />
-
-    </androidx.appcompat.widget.Toolbar>
-
-    <ListView
-        android:id="@+id/list"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        tools:listitem="@layout/logcat_viewer_item_logcat"
-        />
-
-</LinearLayout>
+</androidx.constraintlayout.widget.ConstraintLayout>
